+++ src/main/java/um/tesoreria/report/hexagonal/contable/domain/ports/out/LectivoTotalImputacionRepository.java
@@ -1,0 +1,9 @@
+package um.tesoreria.report.hexagonal.contable.domain.ports.out;
+
+import um.tesoreria.report.hexagonal.contable.domain.model.LectivoTotalImputacion;
+
+import java.util.List;
+
+public interface LectivoTotalImputacionRepository {
+    List<LectivoTotalImputacion> findAllByLectivo(Integer lectivoId);
+}